mkdocs-requirements.txt
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ghp-import==2.1.0
idna==3.7
Jinja2==3.1.3
jsonpath-python==1.0.6
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.19
mkdocs-material-extensions==1.3.1
packaging==24.0
paginate==0.5.6
pathspec==0.12.1
platformdirs==4.2.1
Pygments==2.17.2
pymdown-extensions==10.8
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
urllib3==2.2.1
watchdog==4.0.0
